Separate presentation and business logic from Bookshelf
